requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
jinja2==3.1.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0